Rename service files
