+++ web-app/src/services/teams.ts
@@ -1,25 +1,6 @@
 import _ from 'lodash';
 
-import { checkIfPlayerIsInTeam } from './PlayerService';
-import { TeamInterface, FixtureInterface } from './../interfaces/interfaces';
-
-/**
- * Get search results from a list of team using keyword.
- *
- * @export
- * @param {TeamInterface[]} allTeams
- * @param {string} keyword
- * @returns {TeamInterface[]}
- */
-export function searchTeamsByKeyword(allTeams: TeamInterface[], keyword: string): TeamInterface[] {
-  let teams: TeamInterface[] = [];
-
-  teams = allTeams.filter(team => {
-    return team.name.toLowerCase().indexOf(keyword.toLowerCase()) >= 0 || checkIfPlayerIsInTeam(team, keyword);
-  });
-
-  return teams;
-}
+import { TeamInterface, FixtureInterface } from '@/interfaces/interfaces';
 
 /**
  * Check if the teams is set or to be decided.
@@ -87,6 +68,12 @@ export function isTeamSingle(team: TeamInterface): boolean {
   return !!(team && team.players && team.players.length === 1);
 }
 
+/**
+ * Get custom class names of a team.
+ *
+ * @param team
+ * @returns {object}
+ */
 export function getTeamCustomStyles(team: TeamInterface): object {
   let customStyles = {};
